[REFACTOR]: dto req/res 폴더 분리
