+++ src/main/java/mvp/deplog/domain/auth/application/AuthService.java
@@ -1,8 +1,8 @@
 package mvp.deplog.domain.auth.application;
 
-import mvp.deplog.domain.auth.dto.LoginReq;
-import mvp.deplog.domain.auth.dto.JoinReq;
-import mvp.deplog.domain.auth.dto.LoginRes;
+import mvp.deplog.domain.auth.dto.request.LoginReq;
+import mvp.deplog.domain.auth.dto.request.JoinReq;
+import mvp.deplog.domain.auth.dto.response.LoginRes;
 import mvp.deplog.global.common.Message;
 import mvp.deplog.global.common.SuccessResponse;
