+++ src/main/java/mvp/deplog/domain/auth/dto/request/LoginReq.java
@@ -1,4 +1,4 @@
-package mvp.deplog.domain.auth.dto;
+package mvp.deplog.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
