+++ src/main/java/mvp/deplog/domain/auth/dto/response/LoginRes.java
@@ -1,4 +1,4 @@
-package mvp.deplog.domain.auth.dto;
+package mvp.deplog.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
