+++ src/main/java/mvp/deplog/domain/auth/presentation/AuthController.java
@@ -3,9 +3,9 @@ package mvp.deplog.domain.auth.presentation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import mvp.deplog.domain.auth.application.AuthServiceImpl;
-import mvp.deplog.domain.auth.dto.LoginReq;
-import mvp.deplog.domain.auth.dto.JoinReq;
-import mvp.deplog.domain.auth.dto.LoginRes;
+import mvp.deplog.domain.auth.dto.request.LoginReq;
+import mvp.deplog.domain.auth.dto.request.JoinReq;
+import mvp.deplog.domain.auth.dto.response.LoginRes;
 import mvp.deplog.global.common.Message;
 import mvp.deplog.global.common.SuccessResponse;
 import org.springframework.http.HttpStatus;
