+++ src/main/java/mvp/deplog/domain/post/dto/response/CreatePostRes.java
@@ -1,4 +1,4 @@
-package mvp.deplog.domain.post.dto;
+package mvp.deplog.domain.post.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
