+++ src/main/java/mvp/deplog/domain/post/presentation/PostApi.java
@@ -7,8 +7,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import mvp.deplog.domain.post.dto.CreatePostRes;
-import mvp.deplog.domain.post.dto.PostReq;
+import mvp.deplog.domain.post.dto.response.CreatePostRes;
+import mvp.deplog.domain.post.dto.request.PostReq;
 import mvp.deplog.global.common.Message;
 import mvp.deplog.global.common.SuccessResponse;
 import mvp.deplog.global.exception.ErrorResponse;
